added unit 12

+++ app/Http/Controllers/Admin/PostController.php
@@ -10,6 +10,9 @@ use App\Category;
 use App\Http\Requests\PostStoreFormRequest;
 use App\Http\Requests\PostUpdateFormRequest;
 
+use Gate;
+use Auth;
+
 class PostController extends Controller
 {
     /**
@@ -54,10 +57,21 @@ class PostController extends Controller
      */
     public function create()
     {
-        $categories = Category::all(); 
-        $status = StatusType::toSelectArray(); 
-        $tags = \App\Tag::all();//get()->pluck('name', 'id');
-        return view('admin.posts.create')->withStatus($status)->withCategories($categories)->withTags($tags);
+        $user = \Auth::user();
+        if ($user->can('create', Post::class)) {
+            $categories = Category::all(); 
+            $status = StatusType::toSelectArray(); 
+            $tags = \App\Tag::all();//get()->pluck('name', 'id');
+            return view('admin.posts.create')->withStatus($status)->withCategories($categories)->withTags($tags);
+        } else {
+            return redirect(route('posts.index'))->with('warning','You can not create post');
+        }
+        // if ($this->authorize('create', Post::class)) {
+        //     echo 'Current logged in user is allowed to create new posts.';
+        // } else {
+        //     echo 'You can not create post';
+        // }
+        // exit;
     }
 
     /**
@@ -69,11 +83,13 @@ class PostController extends Controller
     public function store(PostStoreFormRequest $request)
     {
         // Получить post или создать, если не существует...
+        // Get the currently authenticated user's ID...
+        // $id = Auth::id();
         $post = Post::firstOrCreate([
             'title' => $request->title, 
             'content'=>$request->content, 
             'status'=>$request->status, 'category_id'=>$request->category_id, 
-            'user_id'=>1]);
+            'user_id'=>Auth::id()]);
         $post->tags()->sync((array)$request->input('tag'));    
         return redirect(route('posts.index'));
     }
@@ -86,8 +102,17 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        // dump($post);
-        return view('admin.posts.show',compact('post'));
+        $user = \Auth::user();
+        if ($this->authorize('view', $post)) {
+            return view('admin.posts.show',compact('post'));
+        } else {
+            return redirect(route('posts.index'))->with('warning','Not Allowed View Post');
+        }
+        // if ($user->can('view', $post)) {
+        //   echo "Current logged in user is allowed to update the Post: {$post->title}";
+        // } else {
+        //   echo 'Not Authorized.';
+        // }
     }
 
     /**
@@ -98,10 +123,15 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        $categories = Category::pluck('name', 'id'); 
-        $status = StatusType::toSelectArray();
-        $tags = \App\Tag::get()->pluck('name', 'id');
-        return view('admin.posts.edit')->withPost($post)->withStatus($status)->withCategories($categories)->withTags($tags);
+        if (Gate::allows('update-post', $post)) {
+            $categories = Category::pluck('name', 'id'); 
+            $status = StatusType::toSelectArray();
+            $tags = \App\Tag::get()->pluck('name', 'id');
+            return view('admin.posts.edit')->withPost($post)->withStatus($status)->withCategories($categories)->withTags($tags);
+        } else {
+            return redirect(route('posts.index'))->with('warning','Not Allowed Edit Post');
+        }
+        exit;
     }
 
     /**
@@ -113,15 +143,31 @@ class PostController extends Controller
      */
     public function update(PostUpdateFormRequest $request, Post $post)
     {
-        // Если подходящей модели нет, создать новую.
-        $post->updateOrCreate([
-            'title' => $request->title, 
-            'content'=>$request->content, 
-            'status'=>$request->status, 'category_id'=>$request->category_id, 
-            'user_id'=>1
-            ]);
+        // $user = \Auth::user();
+        // if ($user->can('update', $post)) {
+        //     $post->updateOrCreate([
+        //         'title' => $request->title, 
+        //         'content'=>$request->content, 
+        //         'status'=>$request->status, 'category_id'=>$request->category_id, 
+        //         'user_id'=>Auth::id()
+        //         ]);
+        //     $post->tags()->sync((array)$request->input('tag'));
+        //     return redirect(route('posts.index'))->with('success','Post updated successfully');
+        // } else {
+        //     return redirect(route('posts.index'))->with('warning',"Current logged in user is allowed to update the Post: {$post->id}");
+        // }
+        
+        // dd($request);
+        
+        $post->update([
+                'title' => $request->title, 
+                'content'=>$request->content, 
+                'status'=>$request->status, 'category_id'=>$request->category_id, 
+                'user_id'=>Auth::id()
+                ]);
         $post->tags()->sync((array)$request->input('tag'));
-        return redirect(route('posts.index'));
+        return redirect(route('posts.index'))->with('success','Post updated successfully');
+        
     }
 
     /**
@@ -132,9 +178,25 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->tags()->detach();
-        $post->delete();
-        return redirect()->route('posts.index')->with('type','success')->with('message','Post deleted successfully');
+        $user = Auth::user();
+        
+        if ($user->can('delete', $post)) {
+            $post->tags()->detach();
+            $post->delete();
+            return redirect()->route('posts.index')->with('success','Post deleted successfully');
+        } else {
+            return redirect()->route('posts.index')->with('warning','Пользователь '.$user->name.' не может удалять статью...');
+        }
+        
+        // if (Gate::forUser($user)->denies('destroy-post', $post)) {
+        //     // Пользователь не может удалять статью...
+        //     // dd('Пользователь '.$user->name.' не может удалять статью...');
+        //     return redirect()->route('posts.index')->with('warning','Пользователь '.$user->name.' не может удалять статью...');
+        // } else {
+        // $post->tags()->detach();
+        // $post->delete();
+        // return redirect()->route('posts.index')->with('type','success')->with('message','Post deleted successfully');
+        // }
     }
 
     public function getByIds($ids)
